blockchain: Cleanup various tests.
This modifies the newFakeChain and newFakeNode functions to be easier to
work with and then updates the various tests that use them to
significantly simply them.  It also removes the genesisBlockNode
function since it is no longer necessary due to the changes other
changes.

While here, it also improves the comments on the tests to better
describe what is being tested.