blockchain: Modify tests to use chaincfg params.
Rather than redefining the simnet parameters in the tests, just use the
existing chaincfg simnet parameters.  Since some of the tests rely on
the modified legacy version, introduce a clone function for the params
and simply override the parameters that are different.

This is far preferable to redefining them which is cumbersome to update
when the chaincfg params add new definitions.