docs: crosscheck and update all rpc doc entries.
The following updates were done throughout and were made to make important
information in the docs stand out at a glance:

- updates to requests info
- stylistic changes to displaying params
- returns and their types

The extended methods section was also removed (entries are now listed as
standard methods) since that section only applied to upstream. A doc entry for
loadtxfilter was also added.