blockchain.stake: ticket expiry perf improvement.
Change tickettreap to allow for efficient queries by min blockheight,
allowing for O(1) ticket expiry.

This allows for a very significant performance increase as previously
the same query required an iteration of 40,000 or so items while
processing each block.

Remove tickettreap.Mutable - it is not used, and it is difficult to
maintain with the other changes made to the treapNode class used by
Immutable.