coinset: Remove package.
This removes the entire coinset package since it is not used by anything
and, even if something similar to it were desired, it's only related
specifically to wallet functionality in terms of coin selection so
therefore it really should go in the wallet repo.

This is part of an overall effort to rework the layout of the dcrutil
repo to remove unused code and cyclic dependencies.