Import dcrrpcclient repo into rpcclient directory.
This commit contains the entire dcrrpcclient repository along with
several changes needed to move all of the files into the rpcclient
directory in order to prepare it for merging.  This does NOT update dcrd
or any of the other packages to use the new location as that will be
done separately.

- All import paths in the old dcrrpcclient files have been changed to
  the new location
- All references to dcrrpcclient as the package name have been changed to
  rpcclient