sampleconfig: Update min relay fee.
This updates the minimum relay fee in the sample config file to match
the actual default.  Note that it's commented anyways, so this does not
result in any logical change.  However, the file aims to have the
defaults commented out by default, so it makes sense to update the
value accordingly.