stake: Add Hash256PRNG init vector support.
This adds initialization vector support to Hash256PRNG which is used for
determining winning tickets.  This allows callers to cache and reuse the
initialization vector for a given seed to avoid recomputation.

It accomplishes this by exporting two new functions named
CalcHash256PRNGIV and NewHash256PRNGFromIV and updating the existing
NewHash256PRNG function to use the new functions.

It also improves the comment on NewHash256PRNG while here.