blockchain: Rename block index to main chain index.
The current serialized "block index" is actually a main chain index, so
this renames the related functions to correctly denote it's a main chain
index and updates the comments accordingly.

This is being done since future work intends to actually serialize the
entire block index and renaming this now will make the introduction of
full block index code more clear.