blockchain: Limit header sanity check to header.
This modifies the checkBlockHeaderSanity function to accept the header
as intended so that it is only performing context-free checks.  Since
checking the stake difficulty of all of the ticket purchases necessarily
requires the block, that check is moved into checkBlockSanity.

Also, while here, improve the comments on the {c,C}checkProofOfStake
functions.  Note that having comments on an exported function that refer
to an unexported one is not very helpful since documentation generation
tools don't see unexported funcs.