blockchain: No votes/revocations in header sanity.
This changes the validation that the block does not contain any votes or
revocations before stake enabled height and stake validation height to
make use of the values committed to by the header and thus moves the
test into checkBlockHeaderSanity accordingly.

It also combines the two checks to only check before stake validation
height since that always comes after stake enabled height and adds an
assertion to ensure that optimization assumption holds.

This is acceptable since the code also validates the header commitments
for the number of votes and revocations later and therefore implies
correctness.