blockchain: Validate max purchases in header sanity.
This changes the validation that the block does not exceed the maximum
number of ticket purchases to make use of the value committed to by the
header and thus moves the test into checkBlockHeaderSanity accordingly.

This is acceptable since the code also validates the header commitment
for the number of ticket purchases later and therefore implies
correctness.