blockchain: Validate allowed votes in block context.
This moves the test for validating the votes in the block are eligible
votes per the results of the ticket lottery into the checkBlockContext
function where it more naturally belongs since it is only dependent on
the block and its contextual position within the chain.

While here, it also refactors the check into a separate function named
checkAllowedVotes to improve readability.