blockchain: Infrastructure to manage block index.
This adds infrastructure necessary for blockchain to independently
manage and persist a full block index in a bucket.  Note that the
new infrastructure is not used yet as it will be integrated in future
commits after migrating existing databases has been done.

The purpose of this infrastructure is to significantly speed up startup
time which currently has to load a bunch of the most recent blocks in
order to rebuild the stake state and also to pave the way towards being
able to have the full block index in memory allows for a lot of
optimizations and greatly simplified code.

This also includes a full suite of tests to ensure proper serialization
and deserialization functionality.