blockchain: Simplify initial chain state.
Prior to this commit, the chain data structures were initialized in two
separate places depending on whether or not the database state had
already been created.

This simplifies that logic so that the initialization of the data
structures happens in a single place.

In order to accomplish this, the logic has been split into three phases:

1) Query to determine the current state of the database
2) Optionally initialize the database by creating all buckets and
   populating it with the genesis block and associated state if it
   has not already been done
3) Initialize the data structures using the data loaded from the
   database