blockchain: Rework database versioning.
This modifies the database versioning to use separate keys within the
version bucket for each component in order to simplify upgrade logic.

Having a single key for the various versions of the database components
makes it difficult to use because the version isn't known until the data
is read, however, modifying the format necessarily means the version is
needed before that point to know which format to read.

While here, it uses a uint64 to store the date instead of a uint32 so
the timestamp continues to work past 2106.

Also, it removes the upgrade started logic since it should be possible
to stop in the middle of an upgrade and resume later.  Forcing a
completely new database to be created because an upgrade was stopped in
the middle is not ideal.