blockchain: Don't require chain for db upgrades.
This modifies the database upgrade logic to avoid needing the chain and
moves it earlier in the initialization process just after the initial
database creation (if needed), but prior the final struct instance
population from the database.

It is much simpler logic to have an already upgraded database prior to
populating the initial chain state so all of the initial setup happens
in one place.