blockchain: Remove old version information.
This removes the legacy version details from the database to ensure that
loading older versions of the software after the database has been
upgraded fails as desired.

This is important because otherwise, the database could actually be
using a new version that has been upgraded via the new scheme, but old
software would still see this old information and believe the database
was still on that older version.