blockchain: Calculate the lottery IV on demand.
This provides a function to calculate the lottery initialization vector
for a node on demand versus calculating it at node creation time and
storing the result.

This is being done because it saves an additional 32 bytes in the block
node structure and the ultimate goal is to move towards having all block
nodes in memory to greatly simplify and optimize the code.  Since there
will be a lot of nodes in memory, these extra bytes can add up to
significant savings.