blockchain: Refactor block idx entry serialization.
This modifies the newly added block index serialization code to use a
separate structure as opposed to working with block nodes directly.

This approach is more desirable because it provides better separation
and more robust code against changes to the block node structure itself
and it avoids complications dealing with non-serialized fields such as
parent and child pointers.