blockchain: Use chaingen for forced reorg tests.
This replaces and improves the tests for the ForceHeadReorganization
function with new tests based on the dynamic chaingen infrastructure
versus uses static test data that gets out of date as data structures
change and is much more difficult to update.

It also removes the now unused static reorgto180.bz2 test data.