blockchain: Add chain tip tracking.
This adds logic to keep track of all known chain tips.  Currently, this
will only work with chain tips that have been seen since the daemon was
loaded, however, when moving to the full block index in memory, it will
pick end up tracking all chain tips in the block index.

Also, note that nothing makes use of the functionality as of this commit
yet.  It will be useful in the future to allow things such as removal of
the more expensive per-node children tracking, mining code improvements,
and enhanced fork-related interrogation ability.