blockchain: Simplify chain tip tracking.
This simplifies the recently added chain tip tracking by allowing the
block index to handle it when a node is added to it.

This is possible since the block index does not support nodes that do
not connect to an existing node (with the exception of the genesis
block), and thus all new nodes are either extending an existing chain or
are on a side chain, but in either case, are a new chain tip.  In the
case the node is extending a chain, the parent is no longer a tip so it
can be removed.